Refactor: Extract `PrivatAvtaleInnhold` component to streamline rendering logic. Add `formatterSum` utility for improved number formatting. Enhance type separation in `PrivatAvtaleDto`.
