+++ bidrag-dokumentmal/app/routes/bidragskalkulator.privatavtale/route.tsx
@@ -6,30 +6,16 @@ import { ActionFunctionArgs } from "@remix-run/node";
 import HeaderFooter from "~/features/bidragskalkulator/HeaderFooterKalkulator";
 import NavLogo from "~/components/NavLogo";
 import { useActionData } from "@remix-run/react";
-import Innholdsseksjon from "~/features/bidragskalkulator/Innholdsseksjon";
 import {
-  Bidragstype,
-  IPerson,
-  IBarnOgBidrag,
-  IAndreBestemmelser,
-  oppgjørsformTekster,
-  bidragstypeTekster,
   kodeOfNavSkjemaIdKey,
-  Vedleggskrav,
-  vedleggskravTekster,
-  IOppgjør,
   GenererPrivatAvtalePdfRequest,
 } from "~/types/bidragskalkulator";
-import {
-  hentTekst,
-  jaNeiTekster,
-  språkkodeTilSpråkType,
-  SpråkType,
-} from "~/utils/oversettelser";
-import Underskrifter from "~/features/bidragskalkulator/Underskrifter";
+import { hentTekst, språkkodeTilSpråkType } from "~/utils/oversettelser";
+import PrivatAvtaleInnhold from "~/features/bidragskalkulator/PrivatAvtaleInnhold";
 
 // Mock data for development
-const mockRequest: GenererPrivatAvtalePdfRequest = {
+// eslint-disable-next-line unused-imports/no-unused-vars, @typescript-eslint/no-unused-vars
+const mockUnder18Request: GenererPrivatAvtalePdfRequest = {
   privatAvtalePdf: {
     språk: "NB",
     bidragsmottaker: {
@@ -69,6 +55,44 @@ const mockRequest: GenererPrivatAvtalePdfRequest = {
     },
     vedlegg: "INGEN_EKSTRA_DOKUMENTASJON",
   },
+  navSkjemaId: "AVTALE_OM_BARNEBIDRAG_UNDER_18",
+};
+const mockOver18Request: GenererPrivatAvtalePdfRequest = {
+  privatAvtalePdf: {
+    språk: "NB",
+    bidragsmottaker: {
+      fornavn: "Kristian",
+      etternavn: "Etternavnesen",
+      ident: "12345678901",
+    },
+    bidragspliktig: {
+      fornavn: "Kristine",
+      etternavn: "Etternavnesen",
+      ident: "12345678901",
+    },
+    bidrag: [
+      {
+        bidragPerMåned: 2000,
+        fraDato: "Januar, 2025",
+        tilDato: "Februar, 2025",
+      },
+      {
+        bidragPerMåned: 3000,
+        fraDato: "Mars, 2025",
+        tilDato: "April, 2025",
+      },
+    ],
+    oppgjør: {
+      nyAvtale: true,
+      oppgjørsformØnsket: "PRIVAT",
+      oppgjørsformIdag: "PRIVAT",
+    },
+    andreBestemmelser: {
+      harAndreBestemmelser: true,
+      beskrivelse: "Dette er en beskrivelse av andre bestemmelser.",
+    },
+    vedlegg: "INGEN_EKSTRA_DOKUMENTASJON",
+  },
   navSkjemaId: "AVTALE_OM_BARNEBIDRAG_OVER_18",
 };
 
@@ -88,19 +112,19 @@ export async function action(args: ActionFunctionArgs) {
 export default function PrivatAvtaleBidragskalkulator() {
   const actionData = useActionData<{ data: GenererPrivatAvtalePdfRequest }>();
   const response =
-    process.env.NODE_ENV === "development" ? { data: mockRequest } : actionData;
+    process.env.NODE_ENV === "development"
+      ? { data: mockOver18Request }
+      : actionData;
 
   if (response === undefined) {
     return <div>Oops</div>;
   }
 
-  const {
-    data: { privatAvtalePdf, navSkjemaId },
-  } = response;
+  const { privatAvtalePdf, navSkjemaId } = response.data;
+
   const erPrivatAvtaleBarnOver18 =
     navSkjemaId === "AVTALE_OM_BARNEBIDRAG_OVER_18";
   const språk = språkkodeTilSpråkType(privatAvtalePdf.språk) ?? "nb";
-  const innhold = hentTekst(språk, innholdsseksjonTekst);
   const tekster = hentTekst(språk, tekst);
 
   return (
@@ -114,28 +138,7 @@ export default function PrivatAvtaleBidragskalkulator() {
             : tekster.barnUnder18.tittel}
         </h1>
         <p>{kodeOfNavSkjemaIdKey(navSkjemaId)}</p>
-        <div className="flex flex-col gap-4">
-          <Innholdsseksjon
-            tekst={innhold.opplysningerPerson(
-              "MOTTAKER",
-              privatAvtalePdf.bidragsmottaker,
-              true
-            )}
-          />
-          <Innholdsseksjon
-            tekst={innhold.opplysningerPerson(
-              "PLIKTIG",
-              privatAvtalePdf.bidragspliktig
-            )}
-          />
-          <Innholdsseksjon tekst={innhold.barnOgBidrag(privatAvtalePdf.barn)} />
-          <Innholdsseksjon tekst={innhold.oppgjør(privatAvtalePdf.oppgjør)} />
-          <Innholdsseksjon
-            tekst={innhold.andreBestemmelser(privatAvtalePdf.andreBestemmelser)}
-          />
-          <Innholdsseksjon tekst={innhold.vedlegg(privatAvtalePdf.vedlegg)} />
-          <Underskrifter språk={språk} />
-        </div>
+        <PrivatAvtaleInnhold språk={språk} data={response.data} />
       </div>
     </div>
   );
@@ -164,350 +167,3 @@ const tekst = {
     },
   },
 };
-
-const innholdsseksjonTekst = {
-  opplysningerPerson: (
-    bidragstype: Bidragstype,
-    person: IPerson,
-    erOver18: boolean = false
-  ) => {
-    const rolle = bidragstypeTekster[bidragstype];
-    return {
-      nb: {
-        overskrift: `Opplysninger om ${rolle.nb.toLowerCase()} ${erOver18 ? "(barnet over 18 år)" : ""}`,
-        innhold: [
-          { label: "Fornavn", value: person.fornavn, vis: true, type: "text" },
-          {
-            label: "Etternavn",
-            value: person.etternavn,
-            vis: true,
-            type: "text",
-          },
-          ...(person.ident
-            ? [
-                {
-                  label: "Fødselsnummer eller D-nummer (11 siffer)",
-                  value: person.ident,
-                  vis: true,
-                  type: "text",
-                },
-              ]
-            : []),
-        ],
-      },
-      nn: {
-        overskrift: `Opplysningar om ${rolle.nn.toLowerCase()} ${erOver18 ? "(barnet over 18 år)" : ""}`,
-        innhold: [
-          { label: "Fornamn", value: person.fornavn, vis: true, type: "text" },
-          {
-            label: "Etternamn",
-            value: person.etternavn,
-            vis: true,
-            type: "text",
-          },
-          ...(person.ident
-            ? [
-                {
-                  label: "Fødselsnummer eller D-nummer (11 siffer)",
-                  value: person.ident,
-                  vis: true,
-                  type: "text",
-                },
-              ]
-            : []),
-        ],
-      },
-      en: {
-        overskrift: `Information about ${rolle.en.toLowerCase()} ${erOver18 ? "(children over 18 years)" : ""}`,
-        innhold: [
-          {
-            label: "First name",
-            value: person.fornavn,
-            vis: true,
-            type: "text",
-          },
-          {
-            label: "Last name",
-            value: person.etternavn,
-            vis: true,
-            type: "text",
-          },
-          ...(person.ident
-            ? [
-                {
-                  label:
-                    "Norwegian national identification number or D-number (11 digits)",
-                  value: person.ident,
-                  vis: true,
-                  type: "text",
-                },
-              ]
-            : []),
-        ],
-      },
-    };
-  },
-  barnOgBidrag: (barn: IBarnOgBidrag[]) => ({
-    nb: {
-      overskrift: "Opplysninger om barn og bidrag",
-      innhold: barn.flatMap((b, i) => [
-        {
-          label: `Barn ${i + 1}`,
-          value: "",
-          vis: barn.length > 1,
-          type: "heading",
-        },
-        { label: "Fornavn", value: b.fornavn, vis: true, type: "text" },
-        { label: "Etternavn", value: b.etternavn, vis: true, type: "text" },
-        {
-          label: "Fødselsnummer eller D-nummer (11 siffer)",
-          value: b.ident,
-          vis: true,
-          type: "text",
-        },
-        {
-          label: "Beløp per måned",
-          value: `${b.sumBidrag},-`,
-          vis: true,
-          type: "text",
-        },
-        {
-          label: "Avtalen gjelder fra (dd.mm.åååå)",
-          value: b.fraDato,
-          vis: true,
-          type: "text",
-        },
-      ]),
-    },
-    nn: {
-      overskrift: "Opplysningar om barn og bidrag",
-      innhold: barn.flatMap((b, i) => [
-        {
-          label: `Barn ${i + 1}`,
-          value: "",
-          vis: barn.length > 1,
-          type: "heading",
-        },
-        { label: "Fornamn", value: b.fornavn, vis: true, type: "text" },
-        { label: "Etternamn", value: b.etternavn, vis: true, type: "text" },
-        {
-          label: "Fødselsnummer eller D-nummer",
-          value: b.ident,
-          vis: true,
-          type: "text",
-        },
-        {
-          label: "Beløp per månad",
-          value: `${b.sumBidrag},-`,
-          vis: true,
-          type: "text",
-        },
-        {
-          label: "Avtalen gjeld frå (dd.mm.åååå)",
-          value: b.fraDato,
-          vis: true,
-          type: "text",
-        },
-      ]),
-    },
-    en: {
-      overskrift: "Information about child and support",
-      innhold: barn.flatMap((b, i) => [
-        {
-          label: `Child ${i + 1}`,
-          value: "",
-          vis: barn.length > 1,
-          type: "heading",
-        },
-        { label: "First name", value: b.fornavn, vis: true, type: "text" },
-        { label: "Last name", value: b.etternavn, vis: true, type: "text" },
-        {
-          label: "Norwegian national identification number / D number",
-          value: b.ident,
-          vis: true,
-          type: "text",
-        },
-        {
-          label: "Child support amount",
-          value: `${b.sumBidrag},-`,
-          vis: true,
-          type: "text",
-        },
-        {
-          label: "Agreement valid from (dd.mm.yyyy)",
-          value: b.fraDato,
-          vis: true,
-          type: "text",
-        },
-      ]),
-    },
-  }),
-  oppgjør: (oppgjør: IOppgjør) => ({
-    nb: {
-      overskrift: "Oppgjør",
-      innhold: [
-        {
-          label: "Er dette en ny avtale?",
-          value: oppgjør.nyAvtale
-            ? "Ja"
-            : "Nei, dette er en endring av en eksisterende avtale",
-          vis: true,
-          type: "text",
-        },
-        {
-          label: "Hvilken oppgjørsform ønskes?",
-          value: oppgjørsformTekster[oppgjør.oppgjørsformØnsket].nb,
-          vis: true,
-          type: "text",
-        },
-      ],
-    },
-    nn: {
-      overskrift: "Oppgjer",
-      innhold: [
-        {
-          label: "Er dette ein ny avtale?",
-          value: oppgjør.nyAvtale
-            ? "Ja"
-            : "Nei, det er ei endring av eksisterande avtale.",
-          vis: true,
-          type: "text",
-        },
-        {
-          label: "Kva type oppgjer ønskjer de?",
-          value: oppgjørsformTekster[oppgjør.oppgjørsformØnsket].nn,
-          vis: true,
-          type: "text",
-        },
-      ],
-    },
-    en: {
-      overskrift: "Settlement",
-      innhold: [
-        {
-          label: "Is this a new agreement?",
-          value: oppgjør.nyAvtale
-            ? "Yes"
-            : "No, this is an adjustment of an existing agreement.",
-          vis: true,
-          type: "text",
-        },
-        {
-          label: "Which settlement method is desired?",
-          value: oppgjørsformTekster[oppgjør.oppgjørsformØnsket].en,
-          vis: true,
-          type: "text",
-        },
-      ],
-    },
-  }),
-  andreBestemmelser: (andreBestemmelser: IAndreBestemmelser) => ({
-    nb: {
-      overskrift: "Andre bestemmelser",
-      innhold: [
-        {
-          label: "Er det andre bestemmelser tilknyttet avtalen?",
-          value: andreBestemmelser.harAndreBestemmelser
-            ? jaNeiTekster.JA.nb
-            : jaNeiTekster.NEI.nb,
-          vis: true,
-          type: "text",
-        },
-        {
-          label: "Andre bestemmelser tilknyttet avtalen",
-          value: andreBestemmelser.beskrivelse,
-          vis:
-            (andreBestemmelser.harAndreBestemmelser &&
-              andreBestemmelser.beskrivelse) ??
-            false,
-          type: "text",
-        },
-      ],
-    },
-    nn: {
-      overskrift: "Andre bestemmingar",
-      innhold: [
-        {
-          label: "Er det andre bestemmingar knytt til avtalen?",
-          value: andreBestemmelser.harAndreBestemmelser
-            ? jaNeiTekster.JA.nn
-            : jaNeiTekster.NEI.nn,
-          vis: true,
-          type: "text",
-        },
-        {
-          label: "Andre bestemmingar knytt til avtalen",
-          value: andreBestemmelser.harAndreBestemmelser
-            ? jaNeiTekster.JA.en
-            : jaNeiTekster.NEI.en,
-          vis:
-            (andreBestemmelser.harAndreBestemmelser &&
-              andreBestemmelser.beskrivelse) ??
-            false,
-          type: "text",
-        },
-      ],
-    },
-    en: {
-      overskrift: "Other provisions",
-      innhold: [
-        {
-          label: "Are there any other conditions that apply to the agreement?",
-          value: andreBestemmelser.harAndreBestemmelser
-            ? jaNeiTekster.JA.en
-            : jaNeiTekster.JA.en,
-          vis: true,
-          type: "text",
-        },
-        {
-          label: "Other conditions related to the agreement",
-          value: andreBestemmelser.harAndreBestemmelser,
-          vis:
-            (andreBestemmelser.harAndreBestemmelser &&
-              andreBestemmelser.beskrivelse) ??
-            false,
-          type: "text",
-        },
-      ],
-    },
-  }),
-  vedlegg: (vedlegg: Vedleggskrav) => ({
-    nb: {
-      overskrift: "Vedlegg",
-      innhold: [
-        {
-          label: "Har du noen annen dokumentasjon du ønsker å legge ved?",
-          value: vedleggskravTekster[vedlegg].nb,
-          vis: true,
-          type: "text",
-        },
-      ],
-    },
-    nn: {
-      overskrift: "Vedlegg",
-      innhold: [
-        {
-          label: "Annan dokumentasjon",
-          value: vedleggskravTekster[vedlegg].nn,
-          vis: true,
-          type: "text",
-        },
-      ],
-    },
-    en: {
-      overskrift: "Attachments",
-      innhold: [
-        {
-          label: "Other documentation",
-          value: vedleggskravTekster[vedlegg].en,
-          vis: true,
-          type: "text",
-        },
-      ],
-    },
-  }),
-};
-
-export type GeneriskInnholdType = ReturnType<
-  (typeof innholdsseksjonTekst)[keyof typeof innholdsseksjonTekst]
->[SpråkType];
